requirements.txt
asgiref~=3.6.0
Django~=4.2.1
djangorestframework~=3.14.0
simplesrp~=1.0.0b1